Fix #348: fold test-side flock copies into TestFileLock with lock-path parity guard
