+++ previewsmcp/Tests/TestSupportTests/BUILD.bazel
@@ -1,0 +1,13 @@
+load("@rules_swift//swift:swift.bzl", "swift_test")
+
+swift_test(
+    name = "TestSupportTests",
+    srcs = glob(["*.swift"]),
+    copts = [
+        "-swift-version",
+        "6",
+    ],
+    data = ["//tools:simlock"],
+    visibility = ["//visibility:public"],
+    deps = ["//previewsmcp/Tests/TestSupport"],
+)
