+++ tools/BUILD.bazel
@@ -1,0 +1,1 @@
+exports_files(["simlock"])
